pesticide_backend/requirements.txt
-i https://pypi.org/simple
aioredis==1.3.1
asgiref==3.2.10; python_version >= '3.5'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==20.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
autobahn==20.7.1; python_version >= '3.5'
automat==20.2.0
certifi==2020.6.20
cffi==1.14.2
channels-redis==3.0.1
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
daphne==2.5.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-auth==0.9.5
django-richtextfield==1.6
django==3.1
djangorestframework==3.11.1
hiredis==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
hyperlink==20.0.1
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
incremental==17.5.0
msgpack==1.0.0
mysqlclient==2.0.1
pillow==7.2.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyhamcrest==2.0.2; python_version >= '3.5'
pyopenssl==19.1.0
python-decouple==3.3
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
text-unidecode==1.3
twisted[tls]==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
txaio==20.4.1; python_version >= '3.5'
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
zope.interface==5.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
